+++ usrv.c
@@ -47,7 +47,7 @@ static void usrv_msgthr(void *arg)
 
 		switch (oid.id) {
 			case USRV_ID_LOG:
-				log_msgHandler(&msg, oid, rid);
+				log_msgHandler(&msg, oid, (unsigned long)rid);
 				break;
 
 			default:
@@ -68,7 +68,7 @@ void _usrv_start(void)
 	}
 
 	/* MISRA Rule 17.7: Unused returned value, (void) added*/
-	(void)proc_threadCreate(NULL, usrv_msgthr, NULL, 1, SIZE_KSTACK, NULL, 0, NULL);
+	(void)proc_threadCreate(NULL, usrv_msgthr, NULL, 1, (size_t)SIZE_KSTACK, NULL, 0, NULL);
 }
